Refactor governance schema changed notifier
