+++ shardingsphere-governance/shardingsphere-governance-core/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.metadata.schema.refresher.spi.SchemaChangedNotifier
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.driver.governance.internal.metadata.GovernanceSchemaChangedNotifier
+org.apache.shardingsphere.governance.core.event.notifier.GovernanceSchemaChangedNotifier
